Initial commit: move count to aggregation package
