+++ src/query/parser/promql/parse_test.go
@@ -23,6 +23,8 @@ package promql
 import (
 	"testing"
 
+	"github.com/m3db/m3/src/query/functions/aggregation"
+
 	"github.com/m3db/m3/src/query/functions"
 	"github.com/m3db/m3/src/query/functions/binary"
 	"github.com/m3db/m3/src/query/functions/linear"
@@ -43,7 +45,7 @@ func TestDAGWithCountOp(t *testing.T) {
 	assert.Equal(t, transforms[0].Op.OpType(), functions.FetchType)
 	assert.Equal(t, transforms[0].ID, parser.NodeID("0"))
 	assert.Equal(t, transforms[1].ID, parser.NodeID("1"))
-	assert.Equal(t, transforms[1].Op.OpType(), functions.CountType)
+	assert.Equal(t, transforms[1].Op.OpType(), aggregation.CountType)
 	assert.Len(t, edges, 1)
 	assert.Equal(t, edges[0].ParentID, parser.NodeID("0"), "fetch should be the parent")
 	assert.Equal(t, edges[0].ChildID, parser.NodeID("1"), "aggregation should be the child")
