+++ src/query/plan/logical_test.go
@@ -24,6 +24,7 @@ import (
 	"testing"
 
 	"github.com/m3db/m3/src/query/functions"
+	"github.com/m3db/m3/src/query/functions/aggregation"
 	"github.com/m3db/m3/src/query/parser"
 
 	"github.com/stretchr/testify/assert"
@@ -32,7 +33,7 @@ import (
 
 func TestSingleChildParentRelation(t *testing.T) {
 	fetchTransform := parser.NewTransformFromOperation(functions.FetchOp{}, 1)
-	countTransform := parser.NewTransformFromOperation(functions.CountOp{}, 2)
+	countTransform := parser.NewTransformFromOperation(aggregation.CountOp{}, 2)
 	transforms := parser.Nodes{fetchTransform, countTransform}
 	edges := parser.Edges{
 		parser.Edge{
@@ -56,8 +57,8 @@ func TestSingleChildParentRelation(t *testing.T) {
 
 func TestSingleParentMultiChild(t *testing.T) {
 	fetchTransform := parser.NewTransformFromOperation(functions.FetchOp{}, 1)
-	countTransform1 := parser.NewTransformFromOperation(functions.CountOp{}, 2)
-	countTransform2 := parser.NewTransformFromOperation(functions.CountOp{}, 3)
+	countTransform1 := parser.NewTransformFromOperation(aggregation.CountOp{}, 2)
+	countTransform2 := parser.NewTransformFromOperation(aggregation.CountOp{}, 3)
 	transforms := parser.Nodes{fetchTransform, countTransform1, countTransform2}
 	edges := parser.Edges{
 		parser.Edge{
@@ -83,7 +84,7 @@ func TestMultiParent(t *testing.T) {
 	fetchTransform1 := parser.NewTransformFromOperation(functions.FetchOp{}, 1)
 	fetchTransform2 := parser.NewTransformFromOperation(functions.FetchOp{}, 2)
 	// TODO: change this to a real multi parent operation such as asPercent
-	countTransform := parser.NewTransformFromOperation(functions.CountOp{}, 3)
+	countTransform := parser.NewTransformFromOperation(aggregation.CountOp{}, 3)
 
 	transforms := parser.Nodes{fetchTransform1, fetchTransform2, countTransform}
 	edges := parser.Edges{
